change ad ID type to Long

+++ src/main/java/com/yieldlab/biddingservice/controller/BiddingController.java
@@ -23,7 +23,7 @@ public class BiddingController {
     private static final Logger LOGGER = LoggerFactory.getLogger(BiddingController.class);
 
     @GetMapping("{id}")
-    public ResponseEntity<String> initiateAuction(@PathVariable Integer id, @RequestParam Map<String, String> params) {
+    public ResponseEntity<String> initiateAuction(@PathVariable Long id, @RequestParam Map<String, String> params) {
         LOGGER.info("Initiating a new auction for ad ID {} with params: {}", id, params.toString());
         BidResponseDTO auctionWinner = bidsService.holdAuction(id, params);
         String formattedResponse = BidContentFormatter.formatBidContent(auctionWinner);
